Init: made upto GET, POST.
Still working on class BoardDetailView in views.py.
Already possible to upload imagefile from client and other info. like
title, content, boardtype, created_at, updated_at, pk, fk.
The 3 types of boards that we have may not be separated in
pysicalway.The datas of them will be managed at once in a module class
"Board" under the models.py of boards-app.
The practical way to analysis them is that split them as the access of
path from the first click by a client.
The identical path will be given to every acesses through marking
<str:boardtype> and <int:board_id>.

All made by dong. thanks.